Refactor adding main function and modularity
